Identify hetero residues already added to a chain.
Identify hetero residues already added to a chain.  
If a hetero residue id already exists, i.e., a target residue was already added to the current chain, the algorithm checks if this residue has the same name as the already      existing one. If so, all atoms from the target residue will be included as part of the  already existing one, otherwise, a new residue will be created.