Important updates in the interaction calculation methods
InteractionFilter now has a set of properties to control each pairs of molecules should be removed as well as interactions. The control of pairs comprises filters to remove, for example, protein-protein or protein-ligand interactions, etc.

Function to remove orphan H2O pairs, i.e., hydrogen bond involving water molecules and proteins/ligands/etc that does not produce a water-bridged hydrogen bonds, 

New property added to InteractionType: required_interactions. It returns the interactions that one interaction depends on. If no dependent interaction exists, it will return an empty list.

Variable names and packets were updated to recognize the last updates in the directories and file names, as well as, variable names.

Most of the multiple-line commands were updated to fit on only one line