Added new interaction: weak hydrogen bond
A new function called calc_weak_hbond() was added to detect weak hydrogen bonds. The geometrical model for this type of interactions is based on the HBPlus model. It means we have 5 parameters: distances DA and HA, and the angles DHA, D-A-AA, and H-A-AA.

In conf.py, it was added the parameters related to these interactions.

Small changes:
1) Variable names changed to underscore case.
2) Coordinates from coordinate objects are accessed by the property vector: coord.vector.