Amide-interaction working and GroupAtoms with more than one compound
In this update, I finished the tests on amide-pi interaction. Some of these interactions involve two residues forming a peptide bond, what generates an amide group. This group, then, stacks with an aromatic ring. However, so far, my code assumed that each group would be formed by atoms from just one compound. Thus, I had to update the codes to deal with multiple compounds in only one group. Actually, it solved the problem of ligands covalently bonded and I should say that this modeling makes more sense as different groups can be formed when two different ligands bind to each other.

When a group is formed by more than one compound, it will happen that the code will generate two interactions: one for each compound. It happens because each compound will produce the same group, consequently, each group will produce an interaction that should be considered the same. For example, in a peptide bond, two residues R1 and $2 are bond together to form an amide. Departing from R1, we can obtain all of its atom groups, including the amide. The same is true for the R2. To solve it, I had to implement a hash function in some of the classes in order to guarantee that a same set of atoms will be always considered equal.

The updates are:
- coord.py, features.py, groups.py, neighborhood.py, type.py: new functions (or updated functions) __eq__(), __ne__() and __hash__() to guarantee that two coordinates with the same value are considered equal.
- groups.py: function find_compound_groups() addapted to allow the inclusion of multiple residues into only one atom group.
- groups.py: changed function compound() to compounds(). Now an atom group can return all compounds to which its atoms belong to.
- groups.py: new functions is_water(), is_hetatm(), is_aminoacid(), is_nucleotide(), is_mixed(), has_water(), has_hetatm(), has_aminoacid(), has_nucleotide() for checking if the atom group is composed only by or if it contains at least one X (X = water, hetatm, amino acid, nucleotide).
- groups.py: new functions atoms() and a setter for atoms. In the latter case, when atoms is updated, all dependent properties (centroid, coordinates, normal, etc) are also updated.
- calc.py, filter.py, shell_viewer.py: addapted to work with multiple compounds in an atom group.
- conf.py: criteria for amide-pi interactions.
- Residue.py: new function get_class() for get the class (water, hetatm, amino acid, nucleotide, unknown) of a Residue object.

Small updates:
- features.py: camelcase to underscore-case
- features.py: short variable names.
- groups.py: removed function _calc_normal(). The normal is calculated directly in the function normal().