Important update on groups.py
Now, each atom keeps the information of its neighbor and not the neighbor atom per se. The object that keeps the information of the neighbor is NbAtomData (neighborhood.py) that is very similar to the old class Coordinate (coord.py). For that reason, coord.py was removed.

The NbAtomData class stores information about the atomic number,  the coordinates and the serial number of an atom.