Major update: added a new entry type (MolEntry) for managing Mol files
MolEntry allows one to set the molecule that should be analyzed as well as a mol file where the molecule is stored. It is possible to load the file and create a new OBMol object.

I also added a new variable which defines as a regex string how a filename should be. This variable comprises the pattern for evaluating entries whose PDB id is a filename.