Rotating logging files
Main log files created by napoli.py are now stored in a rotating system, i.e., as soon as the file reaches the maximum size (5Mb) a new file is created for logs and the old one is renamed using a identifier that goes from 1 to 10. For example: log, log.1, log.2, etc.