New class for ligand entries definition and new function to recover entries from a PDB
A new class called LigandEntry was created. This class permits one to define a ligand entry both by an object instantiation or from a string.

The new function recover_entries_from_entity() allows one to recover all the entries in a structure/model/chain. It is possible to control if binding ligands and chains are to be searched or not.