Added a new property to the Residue class
It was added a new property called idx to the Residue class. This variable stores the index of a residue in the protein chain sequence, i.e., the order/position in a list formed by the residues of a chain. With such indexes it is possible to select a residue in the residues list without knowing its name or residue number, and without looping over all residues to identify a specific residue.