Fixed bug in function calc_weak_hbond()
Weak hydrogens are calculated by the function calc_weak_hbond() given two specific combination of features: (Acceptor, WeakDonor) or (WeakAcceptor, WeakDonor). 

However, when the code needed to evaluate which groups would be the donor and the acceptor, it was ignoring one of the possible combinations of features (WeakAcceptor and WeakDonor). As a consequence, the code was defining  the donor and acceptor correctly only when the pair of features were Acceptor and WeakDonor.

I also joined the lines of the warnings in this function and added a validation statement to check if both groups contain only one atom.