Fixed another bug in calc_weak_hbond()
Depending on the order that the groups were informed, the code would generate a different result for unfavorable interactions. It was happening especially with groups containing more than one atom passed as the first group and H2O passed as the second group. In these circumstances, the reference used to calculate the NEY angle would be the H2O. Consequently, the code would enter in the loop "if len(dipole_grp2.atoms) == 1" (H2O contain only one atom), producing a valid interaction as no angle could be calculated. However, if you invert the parameter order, the code wouldn't enter in the mentioned loop as the second group would have more than one atom. As a consequence, the code would enter in the Else block where the angle NEY would be calculated. When the angle had been evaluated, it found that it wasn't fit the cutoff rules and the interaction would not be created. To sum up, the interaction would be created only if the H2O is passed as the second group.

So, I fixed this problem by  regulating who will be the first/second dipole when an unfavorable interaction must be calculated. Rule 1:  if a H2O is informed altogether with a dipole with more than one atom, the H2O become the second dipole. Rule 2: if both atoms contain one atom, it doesn't matter who is the first or second group. Rule 3: If both dipoles contain more than one atom, the code evaluates both combinations ((dipole 1, dipole 2) and (dipole 2, dipole 1)). Thus, if both combinations are valid, two interactions will be created.

Besides it, I redefined all log statements inside the functions to be in one line only. And, I added headers over each interaction block in the function _default_functions()