Added directionality to the interactions
Some interactions have directionality. For example, hydrogen bond donors donate their hydrogens to the acceptor. Another example is the Dipole-Dipole interactions where a nucleophile attacks the electrophile. 
To map this characteristic I added a property called directional that works as a flag to set the directionality to On or Off.

So, we changed the name of the properties atm_grp1 and atm_grp2 to src_grp and trgt_grp to characterize the directionality of the interacction, i.e., from the source (src_grp) to the target (trgt_grp). 

Other interactions has this flag set On not because a direction of the interaction as explained before, but because the geometrical parameters related to the interaction depend on the group used as reference. An example of such interactions is the aromatic stacking.