Adapted the code to accept Interactions with direction as its own property
SInce each interaction now has its own property describing its direction, I adapted the code to access this information when the code decides if is necessary to add an arrow or not.

Repulsive interactions are represented in Pymol as two opposite arrows and a block (--<--[]--->--). However, some unfavorable interactions have directionality due to their dependence on the angles calculated. For these cases, the representation will contain only one arrow.