It now consider the number of atoms in the surface into account
If an user wants, it can take into account the number of atoms in the surface and the number of interacting atoms in the surface attends the desired criteria.

The criteria are: min_surf_size and min_inter_atom_in_surf. By default, their values are 1, i.e., only one atom is required in the hydrophobic island.