Added the class AtomGroupNeighborhood to search for contacts between groups
Besides the addition of the class AtomGroupNeighborhood, I also fixed the problem of different order between the MOL and PDBs. The errors were being caused due to a different sorting approach the function _get_all_around_residue() applies and Biopython writter applies. Now the _get_all_around_residue() uses the order in which a residue appears in the Chain list as the Biopython writter.