Fixed error in code related to the full_id size
In to_string() and is_valid() from the class Entry, there was an error due to the full_id size returned by each different entry class. The class MolEntry, for example, returns an array containing only two values and these functions try to access the positions 3 and 4 in the array. So, an error was being raised. Now, it first checks if the array contains more than 2 items before trying to access its third and fourth positions.