Removed the function _merge_hydrophobes().
It won't work with the KDTree, therefore I removed it. Besides it, the identified hydrophobes could be big even if only part of it interacts with the protein. This situation is undesirable as we only want to consider the region that participates to the interaction. A new version is being built after calculating the hydrophobic interactions. The idea of the new approach is to merge only the hydrophobic atoms that establish hydrophobic interaction.