Minor updates: normalization of text and statements, and update of default variables
Besides normalizing the text and statements, and updating the default value of many variables, I also updated the values that are inserted to the interactions property of the Project objects. Now, the tuple to be appended to the interactions list contains an Entry object instead of a PDB file and the calculated interactions to this entry. To make it possible, all entries now sets a PDB file property internally.