Vdw interactions and clashes controlled by two different confs.
The min. number of bonds between two atoms was causing clashes that shouldn't exist to be recognized. If I increase the cutoff, the clash would disappear but the valid hydrogen bonds and other interactions too. So, I decided to create two different variables for controlling the clashes and the other interactions.