Fixed the bug found when selecting a default radius
In the line 375, I was calling the function get_contacts_for_entity passing a radius value that is defined by using the getattr function. The problem  occurs when the property "boundary_cutoff" don't exist in the interaction configuration object. By default the object returns a None value if the property doesn't exist. The expected behavior was that it would end up selecting the default value 6.2 if this property do not exist. So, I fixed it by evaluating if the returned value is None. If so, it gets the default value.