Fixed a bug with the Idx of ligands added with MolEntry
When a new ligand was included in the PDB structure by using a MolEntry object, the ligand was being added without updating its idx. By default, when the ligand object is created in the function get_biopython_structure(), it had its index set to 0 as the ligand structure had only the ligand itself. As a consequence, the ligand is added to the PDB structure with a wrong idx. 

To fix it, I just checked the number of residues/ligands in the chain where the new ligand should be included, and update its idx based on the information already present in the chain.