Removed an inconsistency in multipolar interactions
In some cases, two dipoles with a common atom were being evaluated in this function, which means they are covalently bound. It could happen, for example, in C-S-C as there are two dipoles whose common atom is the S. 

However, it does not make any sense. For that reason, I included a IF statement to check if there is at least one common atom between the dipoles. If so, I return an empty list of interactions.