Solved two bugs: one in PDBs with multiple chains and other related to unexpected addition of hydrogens
Bug 1) The first problem occurred after sorting residues by IDX. The IDX represents the position of a residue in a list stored by its parent chain. As each chain contains its own list, obviously the indexing starts from 0. When I sorted residues only by IDX I didn't realize that it would make residues containing the same IDX would be sorted together. To solved it, I just had to sort residues by its IDX, chain id, and model id.

Bug 2) When the default properties of residues are used, the tool doesn't update or add hydrogens to the protein. However, for ligands, the hydrogen correction is always performed. To do so, we select residues/ligands around a certain radius of a target ligand to guarantee that any compound covalently bound to it are considered during the properties assignment. It occurred that when we do it, we add hydrogens to any molecule recovered during the previous step, which includes residues. After doing this, we update the list of neighbors of an atom according to the recovered list. The problem is that we do it even for residues that should have been treated as DEFAULT or at least equal to the PDB configuration. The problem is even worse because when we select residues around the ligand, we don't recover the potential neighbors of the residue, which makes them AMINO ACIDS and not RESIDUES, and as consequence, they receive a different N protonation. Thus, to solve this problem, the code now only updates the list of neighbors of an atom that belongs to the current compound, i.e., the one to which we need to attribute properties in a specific moment. By doing this, every atom will have its own time to update the list of neighbors while avoiding updating molecules incorrectly.