Fixed bug involving quaternary ammonium N
The validator was only checking for quaternary N containing a degree equal to 5, which happens only when the N has 4 explicit bonds + 1 additional and incorrect bond with an implicit hydrogen, making the N to have valence 5.

However, in amidine N, it does 3 bonds plus the additional invalid bond, which produces a degree equal to 4. In theses cases, the N were not being fixed by the validator.

Now, it fixes  all V5 nitrogens and decides which degree (e.g. 3 or 4) it should use, instead of assuming all Ns to have a degree equal to 4.