Created a new standardization method for residues
The class responsible for the residue standardization is ResiduesStandardiser(). The __init__ of this class accepts a parameter to define if we should break covalent bonds with metals and a histidine protonation state (his_type). However, the last parameter is still not being used so far.

To standardise a molecule, you should call standardise() and provide a list of tuples where the required elements are a PDB atom and a OBAtom. This function comprises all residue atoms and includes rules that detect incorrect valences, charges, aromatic flags, ring flags, bond order, etc. It also provides rules to detect metals and unexpected atoms (other atoms rather than metals or residue atoms) covalently bound to a residue atom. When a metal is bound to a residue, it may try (if it was required by the user) to break the bond and fix the valence, charge, etc, of the atom. After breaking incorrect bonds, free metals are removed and provided as a property called 'removed_atoms' that can be accesses directly through the ResidueStandardiser instance. These covalent bonds with metals are usually incorrectly perceived by Open Babel inasmuch they are actually coordination bonds only.

In the future, I plan to include more filters and standardization rules for common ligands and metals. Another interesting idea is to allow a user to standardize ligands by using a predefined list of SMILES as the ones available for ligands in the RCSB or other molecular databanks.