Updated validator.py to accept Open Babel 3.0
The main important change in Open Babel that impacted our validator method was the remotion of the function SetImplicitValence(), which used to provide a setter for the degree (although the name of the function didn't remind us about its real use). 

Right now, instead of providing a function to define the degree, Open Babel provided a new function called SetImplicitHCount() that allows one to set the number of implicit hydrogens. 

Thus, right now, we need to use SetImplicitHCount() to correct the valence of the atoms.