LUNA projects now can be saved/loaded and compressed if required. Removed user projects code.
Now all save/load functions call pickle_data/unpickle_data from util.file. 

The functions save() and pickle_data() contain a parameter called 'compressed' to inform if the pickled data should be compressed or not. By default, all data are compressed with GZIP.

Moreover, it was removed the user projects code. They should be updated in order to accept the recent updates in the Interaction calculator and in the Atom groups calculator. So, there is no point on keeping these codes in LUNA until it is updated.