FingerprintProject was deprecated
Now to generate fingerprints you need to use LocalProject.generate_fps(),

We also removed the parameters calc_mfp and calc_ifp from the list of properties of a Project instance. Now, you need to use LocalProject.generate_fps() that allows someone to define if MFPs or IFPs should be calculated and if they should be saved in output files.

With this update, now someone is free to change the fingerprint parameters in the Project instance and then regenerate only the necessary fingerprints.