Fixed a bug when loading pickled projects related to circular references
When a project is read from a PKL file, it does not keep the same references for atom groups in AtomGroupsManager and InteractionsManager. Usually, if you update one object in one of the classes, the expected behavior is that the corresponding atom group in the other manager is also updated. However, it does not happen for reloaded projects. Thus, I had to create a code to update references for atom groups just after a project is loaded.

In order to make it work, I also included setters for atom groups in the InteractionType class.