Ligands read from PDB were not being showed in the Pymol session
When a molecule read from a external file (I only found this error with PDB) is loaded into the PDB by using load_mol_from_pdb_block(), it was being added as an additional state in the Pymol session. As a consequence, the molecule was not being showed, but if we changed the state (view of conformations), it appeared.

To fixed it, I set the state of all molecules read from external files to 1. Then everything will be displayed in the same state.