Added a new parameter "is_multimol_file" in the MolEntry
The MolEntry now accepts a parameter called "is_multimol_file" that allows one to set if a mol_file is a multimol file or not. If it is not, the complete file will be read without ensuring that the mol_id corresponds to the id in the file. If it is a multimol, the mol_id should be equal to the one in the file.