Major update: created the PharmacophoreDepiction class
- The function ligand_pharm_figure() from mol.depiction was substituted for the new class PharmacophoreDepiction() that offers better control over Pharmacophore plots. Now, one can also decide if a PNG or SVG must be generated.

- Added a new parameter to the MolEntry class from mol.entry. The parameter 'overwrite_mol_name' is a flag that allows one to define if the original molecule title/name in a molecular file should be substituted by the mol_id informed in a MolEntry instance.

- Updated projects.py with the import of PharmacophoreDepiction, however I still need to update the depiction function to use such class.

Added the functions get_name and set_name in MolWrapper to get or set the name for a molecular object.

Added the function from_smiles in MolWrapper that allows a molecule to be created from smiles.

Added the functions as_rdkit() and as_openbabel() in MolWrapper that converts a molecule to RDKit or Open Babel, respectively.

Updated the color palette for atom types in default_values and included all LUNA types.