Manually increase Recursion limit to 5000
The recursion limit was manually increased to avoid raising the error due to reaching the recursion limit while pickling a project. I found these problems while running a project with Protein-protein interactions. However, this value may not be the best option for general purposes. I need to do more testings