Important update in the logic for selection and expansion of compounds
So far, the strategy for computing pharmacophore properties and invariants worked as follows:

First, molecules nearby a target compound are selected using a boundary cutoff (usually a value as higher as the maximum interaction cutoff).

Then, for each found molecule, it selected compounds in the vicinity using the default covalent radius and only after doing so the properties are computed. By doing so, it tries to correctly assign properties to atoms considering any covalently bound molecule, as otherwise, the results could be fairly incorrect.

For this strategy to work, it assumed that all molecules inside the circle defined by the boundary cutoff would have its own chance to update its atoms' properties. Otherwise, an atom appearing during the covalent selection would have atoms incorrectly perceived. An example is the N atom which could be perceived as being an amine instead of an amide. That happens because the covalent radius only captures molecules directly bound to the central compound, which is the target in a specific moment in the loop.

Thus, this strategy presented problems for molecules in the border of a circle defined by the boundary cutoff as they would never have their own chance to update their atoms' properties as they did not appear  in the mentioned circle.

To solve this problem, any new molecule that appears during the covalent bond selection will be included in the list of molecules that should be processed. Thus, all molecules would have their own chance to update the atoms' properties.

Of course, this new strategy would eventually end up adding all molecules from the protein in the list of selected compounds. To avoid it, we prohibit those compounds found during the expansion to include new molecules into the list of selected compounds.

Latter we also remove any new atom groups not comprising any molecule selected in the original compound list, i.e., only groups comprising the molecules within the circle defined by the boundary cutoff.