Now, all shells are added regardless of their information relevance
I removed the IF statement that used to check if the shell has changed in relation to its previous shell. They are considered different and, therefore, relevant when the shell has expanded its list of atom groups or if any new interactions have been found within the new shell.

However,  I found a bug with patterns identification when using such an approach. 

For a given atom group A1, if its shell at a level L is considered not informative, the program wouldn't include it into the list of shells. As a consequence, in the next level (L+1), when trying to create a new shell, the program would try to access the predecessor level (L) to get the previous shell id, but it wasn't created. So, the program would then try to access an even earlier level (L-2). 

Now, let's say the same atom group A2 in another molecule is able to create its shell in level L. In the next level, differently from what happened for A1,  the program would be able to get the id of this atom group's shell at level L. 

Thus, while A1 uses its shell id in level L-2, A2 uses its id in level L. The problem here is that the hashing value generated will be different just because the program accessed different levels. So, any onwards patterns will be considered different even if they are completely equal.