Major update: MolEntry now implements two constructors
So far, MolEntry used to have a set of parameters that controls two different scenarios. The first is the scenario when the user directly specifies a Mol object, while the other one occurs when the user specifies a Mol file. However, it was creating confusion because some parameters should be defined only in one of these situations. For that reason, I decided to create two constructors which will be described below.

- The __init__() method from MolEntry now accepts only three parameters: the PDB id, the Mol id, and the separator character. The rest of the parameter is initiated with None by default.

- If the user wants to specify a Mol object, he can call the constructor MolEntry.from_mol_obj().  The only mandatory parameters in this scenario are the PDB id, the Mol id, and the Mol object of course.

- On the other hand, if the user wants to specify a file to read the molecule from, he can call MolEntry.from_mol_file.  The only mandatory parameters in this scenario are the PDB id, the Mol id, the Mol file, and the flag is_multimol_file. Note that this flag used to be set to False by default. However, when the user has a multimol file and forgets to change it to True, then the complete file will be read and the first Mol in the file will be used by default. It can completely crash a project. For that reason, now it is mandatory to define if the Mol file is a multimol file or not.

Each constructor provides its specific set of parameters to remove the confusion I mentioned in the first paragraph.