Removed default_properties option because I still need to fix it
- Minor update: all warning calls in the standardiser.py module has been updated to debug.

- Major update: Currently there is a bug in the _set_default_properties() function group groups.py caused by the update in the AtomData class, where I've included a mandatory property called bond_type. 

However, in the _set_default_properties(), I'm trying to create a AtomData object without setting the bond_type. As a consequence, it assumes the serial_number is actually the bond_type and  leaves the serial_number as None. I could easily fix it by placing the property name to obligate the AtomData to recognize the serial number as it is. However, to do so, I would have to make bond_type optional. I don't want to do it  because I use this variable to set covalent bonds, and I want it to stay mandatory.

Thus, since I need to fix it yet, I preferred to remove this function and do not make it usable from now on until I fix _set_default_properties().