Added a new param to InteractionCalculator to control strict rules
It was added the parameter 'lazy_comps_list', which by default receives a list of possible names for Water molecules.

Through this parameter, the user can control which molecules should be skipped from the application of strict rules during Hydrogen or Weak hydrogen bond. Thus, the user can keep the parameters strict_weak_donor_rules and strict_donor_rules active and still apply loose rules in specific molecules.

This is especially useful for solvents like water that usually present more freedom on how to position their hydrogens. Each run of Open Babel, for example, can generate a different position for the hydrogen. As a consequence, the interaction results can contain more interactions depending on where the hydrogen is placed. Given that, the user can specify a list of molecules to skip the strict rules to make the analysis more flexible for solvents, for instance.

When working with MD, however, it is important to consider the hydrogens' position explicitly on every frame. In that sense, the user can set the lazy_comps_list to None and the regular strict/loose rules may be applied.