Fixed a bug in Open Babel caused by an infinite loop
I found a bug in Open Babel while using the entry 1APV:I:DFO:1.

At some point, while converting a molecule, Open Babel enters an infinite loop and for that reason, LUNA never stops running.

To solve this problem, the Popen.communicate() is now called with the parameter timeout=15. So, if Open Babel runs for more than 15s, then the process is killed and it raises a TimeoutExpired error.