Entries containing Open Babel objects are now pickable
As Open Babel does not provide a pickling solution for its objects, I took a walkaround solution. The idea is to convert the mol_obj to a MOL block (yes, strings) and pickle the molecule using that information instead. Then, when the molecule needs to be unpickled, we just need to parse the mol block again. By doing so, the user does not need to provide a molecular object again, or even preprocess it as the molecule will be saved as it is.

This solution was implemented in the MolWrapper class as its own __getstate__ and __setstate__ functions. Also, I've implemented new functions for reading from a MOL block (from_mol_block) or writing a new MOL block (to_mol_block), which are essential to this solution to work.

Thus, in order to make all Entry objects to also be pickable, I just wrap its mol_obj with MolWrapper, when the __getstate__ implemented there will do the work.

Also, in order to take advantage of the MolWrapper functions by default, the mol_obj from the Entry class will be always wrapped when reading it from a file or provided by the user.