Fixed a bug in entry.py, added an updated LUNA environment .yml
In entry.py, when one uses the function recover_entries_from_entity(), if the ligand ID is 0, the entry became an empty space " " due to the way I tested it in the IF statement. Also, when the icode is empty, it was adding an empty space " ", while it should return no character "".

I also added a debug text to alert about entries already processed when the append_mode is used.

In the __init__ from the Project class, there was also a small mistake that it was printing that the inter_calc was not provided even when it did.