Fixed a bug in PDBIO from BioPython
In the current BioPython version, there are two bugs (lines 86 and 95) in the function set_structure() from the class PDBIO (Bio/PDB/PDBIO.py):

       Line 86: sb.structure[0]["A"].add(pdb_object.copy())
       Line 95: sb.structure[0]["A"].child_list[0].add(pdb_object.copy())

As you can see, it always tries to access a chain object whose id is "A". However, if one tries to save a Residue/Atom whose chain id is not "A", it raises a KeyError exception because the recently created chain (line 79) had its id updated to pdb_object.parent.id at line 83 (when pdb_object is a residue) or 92 (when pdb_object is an atom).

To fix this problem, I moved the parent_id to outside the try/except block and substituted the previous lines by

       Line 86: sb.structure[0][parent_id].add(pdb_object.copy())
       Line 95: sb.structure[0][parent_id].child_list[0].add(pdb_object.copy())