Now LUNA can be run sequentially, i.e., not using multiprocessing
When calling LUNA from DASK or from a script that already employs Multiprocessing control, it may raise the error "daemonic processes are not allowed to have children". That happens because the library multiprocessing does not allow daemonic processes to start their own sub-processes.

To solve this problem, one could set the property daemon from multiprocessing.Process to False when creating a pool of workers (be sure to correctly finish your processes in this case), however with DASK it is more complicated.

To solve this problem, I implemented an option where users may run LUNA without using the library multiprocessing, i.e., they will be able to run jobs sequentially.

To do so, when creating a new project be sure to define nproc as NONE.