Fixed bug in loading function and improved the usage of verbosity
The bug I fixed in this version occurs when there is no logging directory in the working path and an user tries to reload a project. When that happens, the program raises an error while trying to access the default logging file. To fix it, it was added a call to prepare_project_path() when the log directory does not exist.

It was also included the properties logging_file and _loaded_logging_file to the Project class. The former defines a default logging file name, while the latter controls if the logging file was already initialized. The property _loaded_logging_file is verified when the verbosity level is updated. If so, it is necessary to update the logging file verbosity level as well.

It was added a getter/setter for the verbosity property, which provides a centralized place for verbosity level validation.