Fixed a bug in RDKit related to dummy atoms / Minor updates at projects.py
In the current version of RDKit, there is a bug caused by some update in RDKit. The error happens when LUNA calls DrawMolecule() after creating dummy atoms and applying SetNoImplicit(True) before adding the atoms to the final molecule.

To fix it, it is necessary to add the dummy atom to the molecule first and only then call  SetNoImplicit.