The function set_view() now highlights all hydrophobic atoms if possible
The function set_view() now tries to recover an AtomGroupManager from one of the interacting atom groups. If so, it will be able to highlight all nearby hydrophobes interacting or not. By doing so, it has the potential to reveal which non-polar atoms are not satisfied.

If the program fails at recovering an AtomGroupsManager, then it will proceed by highlighting only interacting hydrophobes.