Fixed problem when atm_types and feature_extractor are None
When atm_types is None, it was calling the function _perceive_atm_types(). However, if feature_extractor is also None, the code raises an error when it executes the line self.feature_extractor.get_features_by_atoms(rdmol). That happens because a None object does not have the function get_features_by_atoms().

Now, _perceive_atm_types() returns an empty dictionary when feature_extractor is None. In that case, no pharmacophore property will be highlighted on the molecule.