Fixed a bug caused by ligand ids not found in LigandExpo
The bug was found when trying to access the information of a ligand that does not exist in the LigandExpo by calling the property "values" from the returned Pandas DataFrame. In this situation, "values" is an empty array and, therefore, trying to access an index of this array will raise an exception.