Minor update in jobs.py to avoid exceptions
When a processing function is not provided, the saver should try to save the output as it is. To do so, it just needs to access the index 1 of the data object. However, if the returned data is not a list-like object, this access will raise an exception.

For that reason, in this update, I moved the code responsible for getting the output as it is to inside the try-except block.