Fixed a bug related to the TER record and the preserve_atom_numbering
The bug used to happen in the block where the TER record is written when the preserve_atom_numbering flag is set to True. That happened because the serial number was being decremented in one unit. As a consequence, the next atom would have the same serial number as the last atom. Also, this decrementation was actually not necessary as the serial number of a TER record is always one number greater than the serial number of the ATOM/HETATM. Then, the next atom should have a serial number one number greater than the TER record.