Fixed a bug caused by additional lines in a MOL/SDF file before the $$$$
When a MOL/SDF file contains multiple molecules and additional lines before each $$$$ field, the RDKit fails to read the molecule correctly. So, I added a flag that controls if lines after "M  END" and before "$$$$" should be read or not. That solves the problems and avoids reading additional lines not related to the molecule.