Fixed bug caused by atoms with same serial numbers
The bug happened because of the keys used to map neighboring atoms. In the function AtomGroupPerceiver._assign_properties(), I map neighboring atoms using serial numbers, whose information is stored as AtomData objects. Then, this information is accessed in AtomGroupsManager.merge_hydrophobic_atoms() when hydrophobic islands are created. There, I create a new dictionary whose keys are also serial numbers.

The problem with this strategy is that we cannot guarantee that serial numbers are unique. That’s because Biopython only uses the atoms’ full id to guarantee uniqueness. Therefore, serial numbers should not be used as a reference.

This can happen especially when working with PDB files and external molecular files. In these scenarios, external molecular files are converted to PDB format and a new structure object is initialized. As those files don’t have serial number information, Biopython initializes it starting from 1. Consequently, serial numbers in the PDB file may conflict with serial numbers obtained from the external molecular file. That may cause keys in the mapping dictionary to be overwritten with the wrong atom information.

To address this issue, I modified the mentioned functions to work with the atoms’ full id instead of serial numbers.

To make this to work, I had to update very places that use this information to access atomic information.